Giant refactoring update 1
This is a gargantuan update, with too many changes to keep track of. It
is basically my first pass at selectively updating our codebase to have
the latest optimizations and features of bay, /tg/, and paradise. Some
changes worth mentioning off the top of my head:

Under the hood:
-moved controllers into their own processes to be handled via the Goon
processhandler. This should greatly improve performance
-lighting system changed over to object-based lighting
-colored lighting is now possible, and existing light sources have been
updated to give off a light matching their color
-perhaps most importantly, /tg/ garbage collection is now implemented.
This will make things like explosions and Nar sie extremely less laggy.
-moved a bunch of #defines out of global and into subcategories so it's
less of a clusterf***
-breaks air alarms. they will be re-added eventually as we plan on
switching to LINDA for atmospherics anyway.

Mobs:
-introduces some new features from BS12's including limb dislocation and
transplant organ rejection
-organs and grab mechanics are vastly refactored and cleaned up
-surgery refactored
-refactored buckling so that mobs can be buckled to potentially any
object, e.g. handcuffing someone to a pipe or fence. no longer limited
to beds and chairs
-refactored mob resist, adds the container resist proc for things like
getting out of cryo pods
-adds the abilty to stop drop and roll if on fire by using resist
-adds some new mob animations, though not all are implemented yet e.g.
jittery and dizzy
-refactors mob inventory, gone is the u_equip proc. it has been replaced
with unEquip()
-grayscale mob damage overlays which are colored in based on the
species' blood color.
-lots of other smaller refactorings here and there

Items:
-new items added: sushi and fish related item + disco party equipment
from paradise
-desk bell
-judge's gavel  complete with sound fx
-cooking machinery from NT (grill, deep fryer, etc)
-spliffs
-dog beds
-new cooking recipes
-& many more

-refactored duct tape in paperwork & gave it more uses, ex: it can now
be used to muzzle a captive by zone selecting their mouth and attacking
with it
-added duct tape ripping sound effect
-refactored storage-- bags will now display multiples with one icon + a
number overlay instead of repeating them
-ported a detailed fishing system similar to botany from Paradise
-colored light floors give off their respective colored light

Burnable items!
-Certain items can now be burned, indicated by the presence of a
burn_state boolean
-on fire_act() they will catch fire and burn for a configurable amount
of time, giving off a faint glow
-this has its own dedicated process in the processhandler

(Mapping)
-neon signs now have their own objects and lighting
-changed roads from unsimulated to simulated and gave them their own
mapping objects
-some object-less icons that were being used as edited icon_states have
been given their own objects
-colored light floors get their own pre-set items for easy mapping

+++ code/modules/assembly/infrared.dm
@@ -212,12 +212,13 @@
 		return
 	hit()
 
-/obj/effect/beam/i_beam/Del()
-	qdel(next)
-	..()
-	return
-
 /obj/effect/beam/i_beam/Destroy()
-	qdel(next)
+	if(master.first == src)
+		master.first = null
+	if(next)
+		qdel(next)
+		next = null
+	if(previous)
+		previous.next = null
+		master.last = previous
 	..()
-	return
